+++ R/args.R
@@ -44,8 +44,10 @@
 #'   \code{cleanup},
 #'   \code{covariatesTF}, 
 #'   \code{effort},
+#'   \code{ensemble},
 #'   \code{fillweight}, 
 #'   \code{interpolate}, 
+#'   \code{include_interp}, 
 #'   \code{na_drop}, 
 #'   \code{nadot}, 
 #'   \code{NULLname},
@@ -81,10 +83,11 @@
 #'   \code{freq},  
 #'   \code{level}, 
 #'   \code{main}, 
+#'   \code{method} (if not \code{NULL}, must be \code{"unwtavg"}),
 #'   \code{model} (inputted values are checked via 
 #'    \code{\link{verify_models}}), 
 #'   \code{name}, 
-#'   \code{output} (if not \code{NULL}, must be \code{"abundances"}),
+#'   \code{output} (if not \code{NULL}, must be \code{"abundance"}),
 #'   \code{path},
 #'   \code{plots} (if not \code{NULL}, must be \code{"all"} or 
 #'    \code{"longterm"}),
@@ -200,7 +203,7 @@
 #'   \code{ndates} (must be positive),
 #'   \code{nmoons} (must be non-negative),
 #'   \code{start_moon} (must be positive),
-#'   \code{topx} (must be non-negative).
+#'   \code{topx} (must be positive).
 #'
 #'  Must be length-1 \code{integer}-conformable values can be \code{NULL}, 
 #'  and can be \code{NA}:
@@ -209,6 +212,7 @@
 #'
 #'  Must be \code{integer}-conformable values, can be any length, can be 
 #'  \code{NULL}, but cannot be \code{NA}:
+#'   \code{cast_groups} (must be non-negative),
 #'   \code{cast_ids} (must be non-negative),
 #'   \code{end_moons} (must be positive),
 #'   \code{target_moons} (must be positive).
@@ -483,6 +487,7 @@ check_arg_list <- function(){
   }
 
 
+  avail_methods <- c("unwtavg")
   avail_outputs <- c("abundance")
   avail_plots <- c("all", "longterm")
   avail_species <- c("BA", "DM", "DO", "DS", "NA", "NA.", "OL", "OT", "PB", 
@@ -500,6 +505,7 @@ check_arg_list <- function(){
     append_cast_csv = arg_logical(),
     arg_checks = arg_logical(),
     cast = arg_cast(),
+    cast_groups = arg_nonnegintnum(NULL),
     cast_ids = arg_nonnegintnum(NULL),
     cast_id = arg_nonnegintnum(),
     cast_cov = arg_df(),
@@ -534,6 +540,7 @@ check_arg_list <- function(){
     downloads = arg_list(),
     downloads_versions = arg_character(NULL),
     effort = arg_logical(),
+    ensemble = arg_logical(),
     end = arg_date(),
     end_moon = arg_posintnum(),
     end_moons = arg_posintnum(NULL),
@@ -552,6 +559,7 @@ check_arg_list <- function(){
     hist_cov = arg_df(),
     hist_tab = arg_df(),
     interpolate = arg_logical(null = TRUE),
+    include_interp = arg_logical(null = TRUE),
     lag = arg_nonnegintnum(null = TRUE, na = TRUE),
     lat = arg_numeric(),
     lead = arg_posintnum(),
@@ -561,6 +569,7 @@ check_arg_list <- function(){
     lon = arg_numeric(),
     main = arg_character(),
     metadata = arg_list(),
+    method = arg_character(vals = avail_methods),
     min_lag  = arg_nonnegintnum(na = TRUE),
     min_observed = arg_posintnum(),
     min_plots = arg_posintnum(),
@@ -611,6 +620,7 @@ check_arg_list <- function(){
     target_cols = arg_character(NULL),
     time = arg_character(),
     title = arg_character(),
+    topx = arg_posintnum(),
     total = arg_logical(),
     treatment = arg_character(vals = avail_treatments),
     type = arg_character(),