+++ R/process_casts.R
@@ -262,13 +262,11 @@ read_cast_tabs <- function(main = ".", cast_ids = NULL, arg_checks = TRUE){
   }
   cast_tab <- read_cast_tab(main = main, cast_id = cast_ids[1],
                             arg_checks = arg_checks)
-  cast_tab$cast_id <- cast_ids[1]
   ncasts <- length(cast_ids)
   if(ncasts > 1){
     for(i in 2:ncasts){
       cast_tab_i <- read_cast_tab(main = main, cast_id = cast_ids[i],
                                   arg_checks = arg_checks)
-      cast_tab_i$cast_id <- cast_ids[i]
       cast_tab <- rbind(cast_tab, cast_tab_i)
     }
   }
@@ -314,6 +312,10 @@ read_cast_metadata <- function(main = ".", cast_id = NULL, arg_checks = TRUE){
 #'  \code{NULL}, which equates to no selection with respect to 
 #'  \code{end_moon}.
 #'
+#' @param cast_groups \code{integer} (or integer \code{numeric}) value
+#'  of the cast group to combine with an ensemble. If \code{NULL} (default),
+#'  the most recent cast group is ensembled. 
+#'
 #' @param models \code{character} values of the names of the models to 
 #'  include. Default value is \code{NULL}, which equates to no selection with 
 #'  respect to \code{model}.
@@ -324,6 +326,9 @@ read_cast_metadata <- function(main = ".", cast_id = NULL, arg_checks = TRUE){
 #'
 #' @param quiet \code{logical} indicator if progress messages should be
 #'  quieted.
+#'
+#' @param include_interp \code{logical} indicator of if the basic data set
+#'  names should also be inclusive of the associated interpolated data sets.
 #'
 #' @param arg_checks \code{logical} value of if the arguments should be
 #'  checked using standard protocols via \code{\link{check_args}}. The 
@@ -341,9 +346,10 @@ read_cast_metadata <- function(main = ".", cast_id = NULL, arg_checks = TRUE){
 #'
 #' @export
 #'
-select_casts <- function(main = ".", cast_ids = NULL,
+select_casts <- function(main = ".", cast_ids = NULL, cast_groups = NULL,
                          end_moons = NULL, models = NULL, data_sets = NULL,
-                         quiet = FALSE, arg_checks = TRUE){
+                         quiet = FALSE, include_interp = FALSE,
+                         arg_checks = TRUE){
   check_args(arg_checks = arg_checks)
   casts_metadata <- read_casts_metadata(main = main, quiet = quiet, 
                                         arg_checks = arg_checks)
@@ -351,6 +357,10 @@ select_casts <- function(main = ".", cast_ids = NULL,
   ucast_ids <- unique(casts_metadata$cast_id)
   cast_ids <- ifnull(cast_ids, ucast_ids)
   match_id <- casts_metadata$cast_id %in% cast_ids
+
+  ucast_groups <- unique(casts_metadata$cast_group)
+  cast_groups <- ifnull(cast_groups, ucast_groups)
+  match_group <- casts_metadata$cast_group %in% cast_groups
 
   uend_moons <- unique(casts_metadata$end_moon)
   end_moons <- ifnull(end_moons, uend_moons)
@@ -362,6 +372,9 @@ select_casts <- function(main = ".", cast_ids = NULL,
   
   udata_sets <- unique(casts_metadata$data_set)
   data_sets <- ifnull(data_sets, udata_sets)
+  if(include_interp){
+    data_sets <- unique(c(data_sets, paste0(data_sets, "_interp")))
+  }
   match_data_set <- casts_metadata$data_set %in% data_sets
   
   QAQC <- casts_metadata$QAQC
@@ -470,7 +483,11 @@ save_cast_output <- function(cast = NULL, main = ".",
     cast_tab_path <- file_path(main = main, sub = "casts", 
                                files = cast_tab_filename,
                                arg_checks = arg_checks)
-    write.csv(cast$cast_tab, cast_tab_path, row.names = FALSE)
+    cast_tab <- cast$cast_tab
+    cast_tab$cast_id <- next_cast_id
+    cast_tab$cast_group <- cast$metadata$cast_group
+    cast_tab$confidence_level <- cast$metadata$confidence_level
+    write.csv(cast_tab, cast_tab_path, row.names = FALSE)
   }
   if(!is.null(cast$model_fits)){
     model_fits_filename <- paste0("cast_id_", next_cast_id,