+++ tests/testthat/test-15-figures.R
@@ -4,18 +4,19 @@ main <- "./testing"
 
 test_that("plot_cast_point", {
   skip_on_cran() # downloads and casting take too long to run on cran
-  portalcast(main = main, models = c("AutoArima", "NaiveArima"),
-             end_moon = last_moon(main) - 5)
+  portalcast(main = main, 
+             models = c("AutoArima", "NaiveArima"), end_moons = 515:516)
   expect_silent(plot_cast_point(main = main))
+  expect_silent(plot_cast_point(main = main, model = "AutoArima"))
   expect_silent(plot_cast_point(main = main, with_census = TRUE))
-
   expect_error(plot_cast_point(main = main, cast_id = 1e10))
-
+  portalcast(main = main, models = c("AutoArima", "NaiveArima"))
 })
 
 test_that("plot_cast_ts", {
   skip_on_cran() # downloads and casting take too long to run on cran
   expect_silent(plot_cast_ts(main = main))
+  expect_silent(plot_cast_ts(main = main, model = "AutoArima"))
   expect_silent(plot_cast_ts(main = main, species = "DM"))
   expect_error(plot_cast_ts(main = main, cast_id = 1e10))
 
@@ -25,13 +26,16 @@ test_that("plot_cast_ts", {
 test_that("plot_casts_err_lead", {
   skip_on_cran() # downloads and casting take too long to run on cran
   expect_silent(plot_casts_err_lead(main = main))
-  expect_silent(plot_casts_err_lead(main = main, model = "AutoArima", 
-                                   species = "total", data_set = "all"))
-  expect_silent(plot_casts_err_lead(main = main, model = "AutoArima", 
+  expect_silent(plot_casts_err_lead(main = main, models = "AutoArima", 
+                                    ensemble = FALSE, species = "total", 
+                                    data_set = "all"))
+  expect_silent(plot_casts_err_lead(main = main, models = "AutoArima", 
+                                    ensemble = FALSE,
                                    species = "BA", data_set = "all"))
   cast_tab <- read_cast_tabs(main = main)
   expect_error(plot_casts_err_lead(main = main, cast_tab = cast_tab,
                                    cast_id = 1e10))
+  expect_error(plot_casts_err_lead(main = main, cast_id = 1e10))
 })
 
 
@@ -41,4 +45,5 @@ test_that("plot_casts_cov_RMSE", {
   cast_tab <- read_cast_tabs(main = main)
   expect_error(plot_casts_cov_RMSE(main = main, cast_tab = cast_tab,
                                    cast_id = 1e10))
+  expect_error(plot_casts_cov_RMSE(main = main, cast_id = 1e10))
 })